GOt rid of defaultoption class as it was causing bugs
