--- src/main/java/teamthree/twodo/commons/core/options/DefaultOption.java
@@ -1,31 +1,0 @@
-package teamthree.twodo.commons.core.options;
-
-import static teamthree.twodo.commons.util.CollectionUtil.requireAllNonNull;
-
-//@@author A0139267W
-// Sets the default options for 2Do
-public class DefaultOption extends Options {
-
-    private Alarm alarm;
-    private AutoMark autoMark;
-
-    public DefaultOption(Alarm alarm, AutoMark autoMark) {
-        super(alarm, autoMark);
-    }
-
-    public void editAlarm(Alarm newAlarm) {
-        requireAllNonNull(newAlarm);
-        alarm = newAlarm;
-    }
-
-    public void editAutoMark(AutoMark newAutoMark) {
-        requireAllNonNull(newAutoMark);
-        autoMark = newAutoMark;
-    }
-
-    @Override
-    public String toString() {
-        return "\n" + alarm.toString();
-    }
-
-}
